% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R, R/deactivate.R
\name{activate}
\alias{activate}
\alias{active}
\alias{deactivate}
\title{Determine the context of subsequent manipulations.
\code{activate}: To table the focus of future manipulations: elements, sets
or relations.
\code{active}: To check the focus on the \code{TidySet}.
\code{deactivate}: To remove the focus on a specific \code{TidySet}-}
\usage{
activate(.data, what)

active(.data)

deactivate(.data)
}
\arguments{
\item{.data}{A \code{TidySet} object.}

\item{what}{Either "elements", "sets" or "relations"}
}
\value{
A \code{TidySet} object.
}
\description{
Determine the context of subsequent manipulations.
\code{activate}: To table the focus of future manipulations: elements, sets
or relations.
\code{active}: To check the focus on the \code{TidySet}.
\code{deactivate}: To remove the focus on a specific \code{TidySet}-
}
\examples{
relations <- data.frame(sets = c(rep("a", 5), "b", rep("a2", 5), "b2"),
                        elements = rep(letters[seq_len(6)], 2),
                        fuzzy = runif(12))
a <- tidySet(relations)
elements(a) <- cbind(elements(a),
                 type = c(rep("Gene", 4), rep("lncRNA", 2)))
# Filter in the whole TidySet
filter(a, elements == "a")
filter(a, elements == "a", type == "Gene")
# Equivalent to filter_elements
filter_element(a, type == "Gene")
a <- activate(a, "elements")
active(a)
filter(a, type == "Gene")
a <- deactivate(a)
active(a)
filter(a, type == "Gene")
}
