% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/add_column.R
\docType{methods}
\name{add_column}
\alias{add_column}
\alias{add_column,TidySet,character-method}
\title{Add column}
\usage{
add_column(object, slot, columns)

\S4method{add_column}{TidySet,character}(object, slot, columns)
}
\arguments{
\item{object}{A TidySet object.}

\item{slot}{A TidySet slot.}

\item{columns}{The columns to add.}
}
\value{
A \code{TidySet} object.
}
\description{
Add column to a slot of the TidySet object.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = TidySet,slot = character}: Add a column to any slot
}}

\examples{
relations <- data.frame(sets = c(rep("a", 5), "b"),
                        elements = letters[seq_len(6)],
                        fuzzy = runif(6))
a <- tidySet(relations)
add_column(a, "relations", data.frame(well = c("GOOD", "BAD", "WORSE", "UGLY",
"FOE", "HEY")))
}
\seealso{
\code{\link{rename_set}}

Other column: \code{\link{remove_column}}
}
\concept{column}
