% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_elements}
\alias{add_elements}
\title{Add to a TidySet}
\usage{
add_elements(object, elements, ...)
}
\arguments{
\item{object}{A \code{\link{TidySet}} object}

\item{elements}{The new elements}

\item{sets}{The new sets}
}
\value{
A \code{\link{TidySet}} object with the new sets or elements.
}
\description{
Functions to add elements or sets
}
\examples{
x <- list("a" = letters[1:5], "b" = LETTERS[3:7])
a <- tidySet(x)
b <- add_elements(a, "fg")
elements(b)
}
\seealso{
Other add_*: \code{\link{add_sets}}
}
\concept{add_*}
