% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzification.R
\name{add_fuzzification}
\alias{add_fuzzification}
\title{Add a new set}
\usage{
add_fuzzification(object, name, FUN, ...)
}
\arguments{
\item{object}{A TidySet object.}

\item{name}{Name of the set}

\item{FUN}{A function that returns a data.frame}

\item{...}{Arguments passed to FUN}
}
\value{
A TidySet object with the new set.
}
\description{
Given a numeric value and a function adds a new set to the TidySet object.
}
