% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/add_relation.R
\docType{methods}
\name{add_relation}
\alias{add_relation}
\alias{add_relation,TidySet,data.frame-method}
\title{Add relations}
\usage{
add_relation(object, relations, ...)

\S4method{add_relation}{TidySet,data.frame}(object, relations)
}
\arguments{
\item{object}{A TidySet object}

\item{relations}{A data.frame object}

\item{...}{Other arguments.}
}
\value{
A \code{TidySet} object.
}
\description{
Given a TidySets adds relations
}
\section{Methods (by class)}{
\itemize{
\item \code{object = TidySet,relations = data.frame}: Adds relations
}}

\examples{
relations <- data.frame(sets = c(rep("a", 5), "b"),
                        elements = letters[seq_len(6)],
                        fuzzy = runif(6))
a <- tidySet(relations)
relations <- data.frame(sets = c(rep("a2", 5), "b2"),
                        elements = letters[seq_len(6)],
                        fuzzy = runif(6))
add_relation(a, relations)
}
