% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/add_set.R
\docType{methods}
\name{add_set}
\alias{add_set}
\alias{add_set,TidySet,characterORfactor,characterORfactor,numeric-method}
\alias{add_set,TidySet,characterORfactor,characterORfactor,missing-method}
\title{Add a new set}
\usage{
add_set(object, elements, name, fuzzy, ...)


  \S4method{add_set}{TidySet,characterORfactor,characterORfactor,numeric}(object,
  elements, name, fuzzy = 1)


  \S4method{add_set}{TidySet,characterORfactor,characterORfactor,missing}(object,
  elements, name)
}
\arguments{
\item{object}{A TidySet object}

\item{elements}{The elements of the sets.}

\item{name}{The name of the new set.}

\item{fuzzy}{The membership of each element with the set.}

\item{...}{Other arguments.}
}
\value{
A \code{TidySet} object.
}
\description{
Given a TidySets adds a new sets
}
\section{Methods (by class)}{
\itemize{
\item \code{object = TidySet,elements = characterORfactor,name = characterORfactor,fuzzy = numeric}: Adds a set

\item \code{object = TidySet,elements = characterORfactor,name = characterORfactor,fuzzy = missing}: Adds a set
}}

\examples{
relations <- data.frame(sets = c(rep("a", 5), "b"),
                        elements = letters[seq_len(6)],
                        fuzzy = runif(6))
a <- tidySet(relations)
add_set(a, "a", "C")
}
\seealso{
Other add functions: \code{\link{add_relation}}
}
\concept{add functions}
