% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_sets}
\alias{add_sets}
\title{Add to a TidySet}
\usage{
add_sets(object, sets, ...)
}
\arguments{
\item{object}{A \code{\link{TidySet}} object.}

\item{sets}{A character vector with the new sets.}
}
\value{
A \code{\link{TidySet}} object with the new sets.
}
\description{
Functions to add elements or sets
}
\examples{
x <- list("a" = letters[1:5], "b" = LETTERS[3:7])
a <- tidySet(x)
b <- add_sets(a, "fg")
sets(b)
}
\seealso{
Other add_*: \code{\link{add_elements}}
}
\concept{add_*}
