% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{adjacency}
\alias{adjacency}
\title{Adjacency}
\usage{
\method{adjacency}{TidySet}(object)
}
\arguments{
\item{object}{A TidySet object}
}
\value{
A square matrix, 1 if two nodes are connected, 0 otherwise.
}
\description{
Are two elements connected ?
}
\examples{
x <- list("a" = letters[1:5], "b" = LETTERS[3:7])
a <- tidySet(x)
incidence(a)
}
\seealso{
\code{\link{incidence}}
}
