% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange.TidySet}
\alias{arrange.TidySet}
\alias{arrange_set}
\alias{arrange_element}
\alias{arrange_relation}
\title{arrange from a TidySet}
\usage{
\method{arrange}{TidySet}(.data, ...)

arrange_set(.data, ...)

arrange_element(.data, ...)

arrange_relation(.data, ...)
}
\arguments{
\item{.data}{The TidySet object}

\item{...}{Comma separated list of variables names or expressions
integer column position.}
}
\value{
A TidySet object
}
\description{
Use arrange to extract the columns of a TidySet object. You can use activate with filter or
use the specific function. The S3 method filters using all the information
on the TidySet.
}
\examples{
relations <- data.frame(sets = c(rep("a", 5), "b", rep("a2", 5), "b2"),
                        elements = rep(letters[seq_len(6)], 2),
                        fuzzy = runif(12))
a <- tidySet(relations)
a <- mutate_element(a,
                    type = c(rep("Gene", 4), rep("lncRNA", 2)))

b <- arrange(a, desc(type))
elements(b)
b <- arrange_element(a, elements)
elements(b)
# Arrange sets
arrange_set(a, sets)
}
\seealso{
dplyr \code{\link[dplyr]{arrange}} and \code{\link{activate}}
}
