% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame.R
\name{as.data.frame.TidySet}
\alias{as.data.frame.TidySet}
\title{Transforms a TidySet to a data.frame}
\usage{
\method{as.data.frame}{TidySet}(x, ...)
}
\arguments{
\item{x}{The \code{TidySet} object.}

\item{...}{Placeholder for other arguments that could be passed to the
method. Currently not used.}
}
\value{
A \code{data.frame} table.
}
\description{
Flattens the three slots to a single big table
}
