% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{as.list.TidySet}
\alias{as.list.TidySet}
\title{Convert to list}
\usage{
\method{as.list}{TidySet}(x, ...)
}
\arguments{
\item{x}{object to be coerced.}

\item{...}{objects, possibly named (currently ignored).}
}
\value{
A list
}
\description{
Converts a TidySet to a list
}
