% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basesets-package.R
\docType{package}
\name{BaseSet-package}
\alias{BaseSet}
\alias{BaseSet-package}
\title{BaseSet: Provides classes for working with sets}
\description{
What the package does (one paragraph).
}
\details{
It provides a class TidySet with lots of methods to do operations with sets
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/llrs/BaseSet}
  \item Report bugs at \url{https://github.com/llrs/BaseSet/issues}
}

}
\author{
\strong{Maintainer}: Lluís  Revilla Sancho \email{lluis.revilla@gmail.com}

}
\keyword{internal}
