% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basesets-package.R
\docType{package}
\name{BaseSets-package}
\alias{BaseSets-package}
\alias{BaseSets}
\title{BaseSets: A package to handle sets}
\description{
It provides a class TidySet with lots of methods to do operations with sets
}
