% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{combn_indices}
\alias{combn_indices}
\title{Title}
\usage{
combn_indices(x, m)
}
\arguments{
\item{x}{An integer.}

\item{m}{number of elements to choose.}
}
\value{
A list of indices
}
\description{
Title
}
\keyword{internal}
