% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/complement.R
\docType{methods}
\name{complement_element}
\alias{complement_element}
\alias{complement_element,TidySet,characterORfactor-method}
\title{Complement of elements}
\usage{
complement_element(object, elements, ...)

\S4method{complement_element}{TidySet,characterORfactor}(object, elements)
}
\arguments{
\item{object}{A TidySet object.}

\item{elements}{The set to look for the compplement.}

\item{...}{Other arguments.}
}
\value{
A \code{TidySet} object.
}
\description{
Return the objects without the elements listed
}
\section{Methods (by class)}{
\itemize{
\item \code{object = TidySet,elements = characterORfactor}: Complement of the sets.
}}

\examples{
relations <- data.frame(sets = c("a", "a", "b", "b", "c", "c"),
                        elements = letters[seq_len(6)],
                        fuzzy = runif(6))
a <- tidySet(relations)
complement_element(a, "a")
}
\seealso{
Other complements: \code{\link{complement_set}}
}
\concept{complements}
