% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/complement.R
\docType{methods}
\name{complement_set}
\alias{complement_set}
\alias{complement_set,TidySet,character-method}
\title{Complement of a set}
\usage{
complement_set(object, sets, ...)

\S4method{complement_set}{TidySet,character}(object, sets)
}
\arguments{
\item{object}{A TidySet object.}

\item{sets}{The set to look for the compplement.}

\item{...}{Other arguments.}
}
\value{
A \code{TidySet} object.
}
\description{
Return the complement for a set
}
\section{Methods (by class)}{
\itemize{
\item \code{object = TidySet,sets = character}: Complement of the sets.
}}

