% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame.R
\name{df2TS}
\alias{df2TS}
\title{The opposite of as.data.frame}
\usage{
df2TS(.data = NULL, df)
}
\arguments{
\item{.data}{The original TidySet}

\item{df}{The flattened data.frame}
}
\value{
A TidySet object
}
\description{
Convert a data.frame to a TidySet by first using the relations.
It requires the original TidySet in order to convert it back to resemble
the position of the columns.
}
\seealso{
\code{\link[=tidySet.data.frame]{tidySet.data.frame()}}
}
\keyword{internal}
