% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{dimnames.TidySet}
\alias{dimnames.TidySet}
\title{Dimnames of a TidySet}
\usage{
\method{dimnames}{TidySet}(x)
}
\arguments{
\item{x}{A TidySet object.}
}
\value{
A list with the names of the columns of the sets, elements and
relations.
}
\description{
Retrieve the column names of the slots of a TidySet.
}
\examples{
relations <- data.frame(
    sets = c(rep("a", 5), "b"),
    elements = letters[seq_len(6)],
    fuzzy = runif(6)
)
TS <- tidySet(relations)
dimnames(TS)
}
\seealso{
\code{\link[=names]{names()}}
}
