% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplevels.R
\name{droplevels.TidySet}
\alias{droplevels.TidySet}
\title{Drop unused elements and sets}
\usage{
\method{droplevels}{TidySet}(x, elements = TRUE, sets = TRUE, relations = TRUE, ...)
}
\arguments{
\item{x}{A TidySet object.}

\item{elements}{Logical value: Should elements be dropped?}

\item{sets}{Logical value: Should sets be dropped?}

\item{relations}{Logical value: Should sets be dropped?}

\item{...}{Other arguments, currently ignored.}
}
\value{
A TidySet object.
}
\description{
Drop elements and sets without any relation.
}
\examples{
rel <- list(A = letters[1:3], B = character())
TS <- tidySet(rel)
TS
sets(TS)
TS2 <- droplevels(TS)
TS2
sets(TS2)
}
