% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/length.R
\docType{methods}
\name{element_size}
\alias{element_size}
\alias{element_size,TidySet-method}
\title{Calculates the size of the elements}
\usage{
element_size(object, element = NULL)

\S4method{element_size}{TidySet}(object, element = NULL)
}
\arguments{
\item{object}{A TidySet object.}

\item{element}{The element from which the length is calculated.}
}
\value{
A list with the size of the elements or the probability of having that
size.
}
\description{
Calculate the probability of being of different sizes for a given set.
}
\section{Methods (by class)}{
\itemize{
\item \code{TidySet}: Calculates the number of sets one element appears
}}

\seealso{
Other sizes: \code{\link{set_size}}
}
\concept{sizes}
