% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{elements<-}
\alias{elements<-}
\title{Modify the elements}
\usage{
elements(object) <- value
}
\arguments{
\item{object}{Object to be coerced or tested.}

\item{value}{Modification of the elements}
}
\value{
The object with the new elements
}
\description{
Modify the elements
}
