% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/elements.R
\docType{methods}
\name{elements}
\alias{elements}
\alias{elements,TidySet-method}
\title{Retrieve the elements}
\usage{
elements(object)

\S4method{elements}{TidySet}(object)
}
\arguments{
\item{object}{Either a \code{Set} or a \code{SetCollection} object}
}
\description{
Given a Set or a SetCollection retrieve the elements
}
\section{Methods (by class)}{
\itemize{
\item \code{TidySet}: Retrive the elements
}}

