% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{fapply}
\alias{fapply}
\title{Apply to fuzzy}
\usage{
fapply(relations, FUN, ...)
}
\arguments{
\item{relations}{A data.frame or similar with fuzzy, sets and elements
columns.}

\item{FUN}{A function to perform on the fuzzy numbers.}

\item{...}{Other named arguments passed to \code{FUN}.}
}
\value{
A modified TidySet object
}
\description{
Simplify and returns unique results of the object.
}
\keyword{internal}
