% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\alias{filter_set}
\alias{filter_element}
\alias{filter_relation}
\alias{filter.TidySet}
\title{Filter}
\usage{
filter(.data, ...)

filter_set(.data, ...)

filter_element(.data, ...)

filter_relation(.data, ...)

\method{filter}{TidySet}(.data, ...)
}
\arguments{
\item{.data}{The TidySet object}

\item{...}{The logical predicates in terms of the variables of the sets}
}
\value{
A TidySet object
}
\description{
Use filter to subset the TidySet object. You can use activate with filter or
use the specific function. The S3 method filters using all the information
on the TidySet.
}
\examples{
relations <- data.frame(sets = c(rep("a", 5), "b", rep("a2", 5), "b2"),
                        elements = rep(letters[seq_len(6)], 2),
                        fuzzy = runif(12))
a <- tidySet(relations)
elements(a) <- cbind(elements(a),
                 type = c(rep("Gene", 4), rep("lncRNA", 2)))
filter(a, elements == "a")
# Equivalent to filter_relation
filter(a, elements == "a", sets == "a")
filter_relation(a, elements == "a", sets == "a")
# Filter element
filter_element(a, type == "Gene")
# Filter sets and by property of elements simultaneously
filter(a, sets == "b", type == "lncRNA")
# Filter sets
filter_set(a, sets == "b")
}
\seealso{
dplyr \code{\link[dplyr]{filter}} and \code{\link{activate}}
}
