% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzification.R
\name{fuzzification}
\alias{fuzzification}
\title{Transform a numeric value to a TidySet}
\usage{
fuzzification(name, FUN, ...)
}
\arguments{
\item{name}{Name of the set}

\item{FUN}{A function that returns a data.frame}

\item{...}{Arguments passed to FUN}
}
\value{
A TidySet object.
}
\description{
Transform a numeric value to a TidySet
}
