% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMT.R
\name{getGMT}
\alias{getGMT}
\title{Import GMT files}
\usage{
getGMT(con, sep = "\\t", ...)
}
\arguments{
\item{con}{File name of the GMT file.}

\item{sep}{Separation of the file}

\item{...}{Other arguments passed to strsplit}
}
\value{
A TidySet object
}
\description{
Import GMT files
}
\examples{
gmtText <- c(
"KEGG_GLYCOLYSIS_GLUCONEOGENESIS\\thttp://www.broadinstitute.org/gsea/msigdb/cards/KEGG_GLYCOLYSIS_GLUCONEOGENESIS\\tACSS2\\tGCK\\tPGK2\\tPGK1",
"KEGG_CITRATE_CYCLE_TCA_CYCLE\\thttp://www.broadinstitute.org/gsea/msigdb/cards/KEGG_CITRATE_CYCLE_TCA_CYCLE\\tIDH3B\\tDLST\\tPCK2" )
gs <- getGMT(textConnection(gmtText))
gs
}
