% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obo.R
\name{getOBO}
\alias{getOBO}
\title{Read an OBO file}
\usage{
getOBO(x)
}
\arguments{
\item{x}{A file in OBO format}
}
\value{
A TidySet object
}
\description{
Read a OBO formatted file
}
\examples{
oboFile <- system.file(package = "BaseSet", "extdata",
                                  "go-basic_subset.obo")
gs <- getOBO(oboFile)
head(gs)
}
\seealso{
Other IO functions: 
\code{\link{getGAF}()},
\code{\link{getGMT}()},
\code{\link{getXML}()}
}
\concept{IO functions}
