% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{getXML}
\alias{getXML}
\title{Read an XML file}
\usage{
getXML(x)
}
\arguments{
\item{x}{A file in XML format}
}
\value{
A TidySet object
}
\description{
Read an XML formatted file
}
\seealso{
Other IO functions: 
\code{\link{getGAF}()},
\code{\link{getGMT}()},
\code{\link{getOBO}()}
}
\concept{IO functions}
