% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group}
\alias{group}
\alias{group.TidySet}
\title{Create a new set from existing elements}
\usage{
group(object, name, ...)

\method{group}{TidySet}(object, name, ...)
}
\arguments{
\item{object}{A TidySet object.}

\item{name}{The name of the new set.}

\item{...}{A logical condition to subset some elements}
}
\value{
A TidySet object with the new set
}
\description{
Create a new set from existing elements
}
