% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/incidence.R
\docType{methods}
\name{incidence}
\alias{incidence}
\alias{incidence,TidySet-method}
\title{Incidence}
\usage{
incidence(object)

\S4method{incidence}{TidySet}(object)
}
\arguments{
\item{object}{Object to be coerced or tested.}
}
\description{
Check which elements are in which sets.
}
\section{Methods (by class)}{
\itemize{
\item \code{TidySet}: Incidence of the TidySet
}}

\examples{
x <- list("a" = letters[1:5], "b" = LETTERS[3:7])
a <- tidySet(x)
incidence(a)
}
