% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set.R, R/setcollection.R
\docType{methods}
\name{is.fuzzy}
\alias{is.fuzzy}
\alias{is.fuzzy,Set-method}
\alias{is.fuzzy,SetCollection-method}
\title{Fuzzy Sets}
\usage{
is.fuzzy(object)

\S4method{is.fuzzy}{Set}(object)

\S4method{is.fuzzy}{SetCollection}(object)
}
\arguments{
\item{object}{Object to be coerced or tested.}
}
\description{
Check if there are fuzzy sets.
}
\section{Methods (by class)}{
\itemize{
\item \code{Set}: For \code{Set} objects

\item \code{SetCollection}: For \code{SetCollection} objects
}}

