% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{length.TidySet}
\alias{length.TidySet}
\title{Length of the TidySet}
\usage{
\method{length}{TidySet}(x)
}
\arguments{
\item{x}{A TidySet object.

No replacement function is available, either delete sets or add them.}
}
\value{
A numeric value.
}
\description{
Returns the number of sets in the object.
}
\examples{
TS <- tidySet(list(A = letters[1:5], B = letters[6]))
length(TS)
}
\seealso{
\code{\link[=dim]{dim()}}, \code{\link[=ncol]{ncol()}} and \code{\link[=nrow]{nrow()}}.
Also look at \code{\link[=lengths]{lengths()}} for the number of relations of sets.
}
