% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{length_helper}
\alias{length_helper}
\title{Calculates the probabilities}
\usage{
length_helper(p, i)
}
\arguments{
\item{p}{Probabilities}

\item{i}{Index of the complemetary probability}
}
\description{
Calculates the probability that all happened simultaneously
}
\keyword{internal}
