% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{length_set}
\alias{length_set}
\title{Calculates the probabil}
\usage{
length_set(fuzziness)
}
\arguments{
\item{fuzziness}{The probabilities}
}
\value{
A vector with the probability of each set
}
\description{
Calculates the probabil
}
\examples{
length_set(c(0.5, 0.1, 0.3, 0.5, 0.25, 0.23))
}
