% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{lengths,TidySet-method}
\alias{lengths,TidySet-method}
\title{Lengths of the TidySet}
\usage{
\S4method{lengths}{TidySet}(x, use.names = TRUE)
}
\arguments{
\item{x}{A TidySet object.}

\item{use.names}{A logical value whether to inherit names or not.}
}
\value{
A vector with the number of different relations for each set.
}
\description{
Returns the number of relations of each set in the object.
}
\examples{
TS <- tidySet(list(A = letters[1:5], B = letters[6]))
lengths(TS)
}
\seealso{
\code{\link[=length]{length()}}, Use \code{\link[=set_size]{set_size()}} if you are using fuzzy sets.
}
