% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/move_to.R
\docType{methods}
\name{move_to}
\alias{move_to}
\alias{move_to,TidySet,characterORfactor,characterORfactor,character-method}
\title{Move columns}
\usage{
move_to(object, from, to, columns)


  \S4method{move_to}{TidySet,characterORfactor,characterORfactor,character}(object,
  from, to, columns)
}
\arguments{
\item{object}{A TidySet object.}

\item{from}{The name of the slot where the content is.}

\item{to}{The name of the slot to move the content.}

\item{columns}{The name of the columns that should be moved.}
}
\value{
A TidySet object where the content is moved from one slot to other.
}
\description{
Moves information from one slot to other slots.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = TidySet,from = characterORfactor,to = characterORfactor,columns = character}: Move columns
}}

\examples{
x <- list("a" = c("A" = 0.1, "B" = 0.5), "b" = c("A" = 0.2, "B" = 1))
a <- tidySet(x)
a <- mutate_element(a, b = runif(2))
b <- move_to(a, from = "elements", to = "relations", "b")
}
