% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate.TidySet}
\alias{mutate.TidySet}
\alias{mutate_set}
\alias{mutate_element}
\alias{mutate_relation}
\title{Mutate}
\usage{
\method{mutate}{TidySet}(.data, ...)

mutate_set(.data, ...)

mutate_element(.data, ...)

mutate_relation(.data, ...)
}
\arguments{
\item{.data}{The TidySet object}

\item{...}{The logical predicates in terms of the variables of the sets}
}
\value{
A TidySet object
}
\description{
Use mutate to alter the TidySet object. You can use activate with mutate or
use the specific function. The S3 method filters using all the information
on the TidySet.
}
\examples{
relations <- data.frame(sets = c(rep("a", 5), "b", rep("a2", 5), "b2"),
                        elements = rep(letters[seq_len(6)], 2),
                        fuzzy = runif(12))
a <- tidySet(relations)
a <- mutate_element(a, Type = c(rep("Gene", 4), rep("lncRNA", 2)))
a
b <- mutate_relation(a, Type = sample(c("PPI", "PF", "MP"), 12,  replace = TRUE))
}
\seealso{
\code{\link[dplyr]{filter}} and \code{\link{activate}}
}
