% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{name_elements<-}
\alias{name_elements<-}
\title{Rename elements}
\usage{
name_elements(object) <- value
}
\arguments{
\item{object}{A TidySet object.}

\item{value}{A character with the new names for the elements.}
}
\value{
A \code{TidySet} object.
}
\description{
Rename elements.
}
\examples{
relations <- data.frame(sets = c(rep("a", 5), "b"),
                        elements = letters[seq_len(6)],
                        fuzzy = runif(6))
a <- tidySet(relations)
a
name_elements(a) <- letters[1:6]
}
\seealso{
\code{\link{rename_elements}}

Other names: \code{\link{name_elements}},
  \code{\link{name_sets<-}}, \code{\link{name_sets}},
  \code{\link{rename_elements}}, \code{\link{rename_set}}
}
\concept{names}
