% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/names.R
\docType{methods}
\name{name_elements}
\alias{name_elements}
\alias{name_elements,TidySet-method}
\title{Name elements}
\usage{
name_elements(object)

\S4method{name_elements}{TidySet}(object)
}
\arguments{
\item{object}{A TidySet object}
}
\value{
A \code{TidySet} object
}
\description{
Retrieve the name of the elements
}
\section{Methods (by class)}{
\itemize{
\item \code{TidySet}: Name elements
}}

\seealso{
Other name: \code{\link{name_sets}}
}
\concept{name}
