% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{name_sets<-}
\alias{name_sets<-}
\title{Rename sets}
\usage{
name_sets(object) <- value
}
\arguments{
\item{object}{A TidySet object.}

\item{value}{A character with the new names for the sets.}
}
\value{
A \code{TidySet} object.
}
\description{
Rename sets.
}
\examples{
relations <- data.frame(sets = c(rep("a", 5), "b"),
                        elements = letters[seq_len(6)],
                        fuzzy = runif(6))
a <- tidySet(relations)
a
name_sets(a) <- LETTERS[1:2]
}
\seealso{
\code{\link{rename_set}}

Other names: \code{\link{name_elements<-}}
}
\concept{names}
