% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/names.R
\docType{methods}
\name{name_sets}
\alias{name_sets}
\alias{name_sets,TidySet-method}
\title{Name sets}
\usage{
name_sets(object)

\S4method{name_sets}{TidySet}(object)
}
\arguments{
\item{object}{A TidySet object}
}
\value{
A \code{TidySet} object
}
\description{
Retrieve the name of the sets
}
\section{Methods (by class)}{
\itemize{
\item \code{TidySet}: Name sets
}}

\concept{name}
