% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{nRelations}
\alias{nRelations}
\title{Number of relations}
\usage{
nRelations(object)
}
\arguments{
\item{object}{Object to be coerced or tested.}
}
\description{
Check the number of relations of the TidySet
}
