% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{relation}
\alias{relation}
\title{Find strength of the relationship}
\usage{
relation(object)
}
\arguments{
\item{object}{Object to be coerced or tested.}
}
\value{
A list of weight with the relationship. If it is not a fuzzy set
it is converted to 1.
}
\description{
Transform the object to a tidy/long \code{data.frame}
}
