% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{relations<-}
\alias{relations<-}
\title{Modify the relationships}
\usage{
relations(object) <- value
}
\arguments{
\item{object}{Object to be coerced or tested.}

\item{value}{Modification of the relationsihps}
}
\value{
The object with the new relationships
}
\description{
Modify the relationships
}
