% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/relations.R
\docType{methods}
\name{relations}
\alias{relations}
\alias{relations<-}
\alias{relations,TidySet-method}
\alias{relations<-,TidySet-method}
\title{Find the relationship}
\usage{
relations(object)

\S4method{relations}{TidySet}(object)

\S4method{relations}{TidySet}(object) <- value
}
\arguments{
\item{object}{Object to be coerced or tested.}

\item{value}{Modification of the relationsihps}
}
\value{
The existing relationships
}
\description{
Find the relationship
}
\section{Methods (by class)}{
\itemize{
\item \code{TidySet}: Retrieve the relations

\item \code{TidySet}: Modify the relations
}}

\seealso{
\code{\link{nRelations}}

Other slots: \code{\link{elements}}, \code{\link{sets}}
}
\concept{slots}
