% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/remove_column.R
\docType{methods}
\name{remove_column}
\alias{remove_column}
\alias{remove_column,TidySet,character,character-method}
\title{Remove column}
\usage{
remove_column(object, slot, column_names)

\S4method{remove_column}{TidySet,character,character}(object, slot,
  column_names)
}
\arguments{
\item{object}{A TidySet object}

\item{slot}{A TidySet slot}

\item{column_names}{The name of the columns}
}
\value{
A \code{TidySet} object
}
\description{
Removes column from a slot of the TidySet object
}
\section{Methods (by class)}{
\itemize{
\item \code{object = TidySet,slot = character,column_names = character}: Remove columns to any slot
}}

\seealso{
\code{\link{rename_set}}

Other column: \code{\link{add_column}}
}
\concept{column}
