% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/remove.R
\docType{methods}
\name{remove_element}
\alias{remove_element}
\alias{remove_element,TidySet,characterORfactor-method}
\title{Remove elements}
\usage{
remove_element(object, elements, ...)

\S4method{remove_element}{TidySet,characterORfactor}(object, elements)
}
\arguments{
\item{object}{A TidySet object.}

\item{elements}{The elements to be removed.}

\item{...}{Other arguments.}
}
\value{
A \code{TidySet} object.
}
\description{
Given a TidySets remove elements and the related relations and if
required also the sets.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = TidySet,elements = characterORfactor}: Removes everything related to an element
}}

\examples{
relations <- data.frame(sets = c(rep("a", 5), "b"),
                        elements = letters[seq_len(6)],
                        fuzzy = runif(6))
a <- tidySet(relations)
remove_element(a, "c")
}
