% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/remove.R
\docType{methods}
\name{remove_set}
\alias{remove_set}
\alias{remove_set,TidySet,character,character-method}
\title{Remove a set}
\usage{
remove_set(object, elements, sets, ...)

\S4method{remove_set}{TidySet,character,character}(object, elements, sets)
}
\arguments{
\item{object}{A TidySet object}

\item{elements}{The elements of the sets.}

\item{...}{Other arguments}

\item{setName}{The name of the new set.}

\item{fuzzy}{The membership of each element with the set.}
}
\value{
A \code{TidySet} object
}
\description{
Given a TidySets remove a set
}
\section{Methods (by class)}{
\itemize{
\item \code{object = TidySet,elements = character,sets = character}: Removes a set
}}

