% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/rename.R
\docType{methods}
\name{rename_elements}
\alias{rename_elements}
\alias{rename_elements,TidySet-method}
\title{Rename elements}
\usage{
rename_elements(object, old, new)

\S4method{rename_elements}{TidySet}(object, old, new)
}
\arguments{
\item{object}{A TidySet object.}

\item{old}{A character vector of to be renamed.}

\item{new}{A character vector of with new names.}
}
\value{
A \code{TidySet} object.
}
\description{
Change the default names of sets and elements.
}
\section{Methods (by class)}{
\itemize{
\item \code{TidySet}: Rename elements
}}

\seealso{
\code{\link{name_elements}}

Other renames: \code{\link{rename_set}}

Other names: \code{\link{name_elements}},
  \code{\link{name_sets}}, \code{\link{rename_set}}
}
\concept{names}
\concept{renames}
