% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.TidySet}
\alias{select.TidySet}
\alias{select_set}
\alias{select_element}
\alias{select_relation}
\title{select from a TidySet}
\usage{
\method{select}{TidySet}(.data, ...)

select_set(.data, ...)

select_element(.data, ...)

select_relation(.data, ...)
}
\arguments{
\item{.data}{The TidySet object}

\item{...}{The name of the columns you want to keep, remove or rename.}
}
\value{
A TidySet object
}
\description{
Use select to extract the columns of a TidySet object. You can use activate with filter or
use the specific function. The S3 method filters using all the information
on the TidySet.
}
\examples{
relations <- data.frame(sets = c(rep("a", 5), "b", rep("a2", 5), "b2"),
                        elements = rep(letters[seq_len(6)], 2),
                        fuzzy = runif(12))
a <- tidySet(relations)
elements(a) <- cbind(elements(a),
                 type = c(rep("Gene", 4), rep("lncRNA", 2)))

b <- select(a, -type)
elements(b)
b <- select_element(a, elements)
elements(b)
# Filter sets
select_set(a, sets)
}
\seealso{
dplyr \code{\link[dplyr]{select}} and \code{\link{activate}}
}
