% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set}
\alias{set}
\title{Create a set}
\usage{
set(elements)
}
\arguments{
\item{elements}{Elements of the set}
}
\value{
An object of class Set
}
\description{
Create a set
}
\examples{
a <- set(c("a", "b"))
e <- set(c("a" = 0.1, "b" = 0.5))
}
