% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\docType{methods}
\name{set_size,TidySet-method}
\alias{set_size,TidySet-method}
\title{Set size}
\usage{
\S4method{set_size}{TidySet}(object, set)
}
\arguments{
\item{set}{A set of class \code{Set}}
}
\value{
A vector with the length of the set and its probability
}
\description{
Calculates the size of a set either fuzzy or not
}
