% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{set_size}
\alias{set_size}
\title{Set size}
\usage{
set_size(set)
}
\arguments{
\item{set}{A set of class \code{Set}}
}
\value{
A vector with the length of the set and its probability
}
\description{
Calculates the size of a set either fuzzy or not
}
