% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{set_symbols}
\alias{set_symbols}
\title{A subset of symbols related to sets}
\format{
An object of class \code{character} of length 16.
}
\usage{
set_symbols
}
\description{
Name and symbol of operations related to sets, including intersection
and union among others:
}
\examples{
set_symbols
}
\references{
\url{https://www.fileformat.info/info/unicode/category/Sm/list.htm}
}
\keyword{datasets}
