% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SetCollection-class}
\alias{SetCollection-class}
\title{A class to represent a group of sets}
\description{
Several sets together form a collection
}
\section{Slots}{

\describe{
\item{\code{sets}}{A group of unique sets}
}}

