% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{setCollection}
\alias{setCollection}
\title{Create a SetCollection}
\usage{
setCollection(sets)
}
\arguments{
\item{sets}{The list of sets}
}
\value{
An object of class SetCollection
}
\description{
Create a SetCollection
}
\examples{
a <- set(c("a", "b"))
b <- set(c("a", "b"))
setCollection(c(a, b))
}
