% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{sets<-}
\alias{sets<-}
\title{Modify the sets}
\usage{
sets(object) <- value
}
\arguments{
\item{object}{Object to be coerced or tested.}

\item{value}{Modification of the sets}
}
\value{
The object with the new sets
}
\description{
Modify the sets
}
