% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{sets}
\alias{sets}
\title{Retrieve the names of the sets}
\usage{
sets(object)
}
\arguments{
\item{object}{A \code{SetCollection} object}
}
\description{
Given a SetCollection retrieve the names of the sets
}
