% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{show,TidySet-method}
\alias{show,TidySet-method}
\title{Method to show the TidySet object}
\usage{
\S4method{show}{TidySet}(object)
}
\arguments{
\item{object}{A TidySet}
}
\value{
A table with the information of the relationships.
}
\description{
Prints the resulting table of a TidySet object. Does not shown elements or
sets without any relationship (empty sets). To see them use \code{\link[=sets]{sets()}} or
\code{\link[=elements]{elements()}}.
}
