% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setcollection.R
\docType{methods}
\name{[,SetCollection-method}
\alias{[,SetCollection-method}
\title{Extract or Replace Parts of a SetCollection}
\usage{
\S4method{[}{SetCollection}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{SetCollection from which to extract element(s) or in which to
replace element(s)}

\item{i, j}{indices specifying elements to extract or replace. Indices are
numeric or character vectors or empty (missing) or \code{NULL}}

\item{...}{Other arguments ignored}

\item{drop}{Argument ignored}
}
\description{
Equivalent of the base [], and [[]]
}
