% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/subtract.R
\docType{methods}
\name{subtract}
\alias{subtract}
\alias{subtract,TidySet,characterORfactor,characterORfactor-method}
\title{Subtract}
\usage{
subtract(object, set_in, not_in, ...)

\S4method{subtract}{TidySet,characterORfactor,characterORfactor}(object,
  set_in, not_in, name = NULL, keep = TRUE, keep_relations = keep,
  keep_elements = keep, keep_sets = keep)
}
\arguments{
\item{object}{A TidySet object.}

\item{set_in}{Name of the sets where the elements should be present.}

\item{not_in}{Name of the sets where the elements should not be present.}

\item{...}{Other arguments.}

\item{name}{Name of the new set. By default it adds a "C_".}

\item{keep}{Logical value to keep all the other sets.}

\item{keep_relations}{A logical value if you wan to keep old relations}

\item{keep_elements}{A logical value if you wan to keep old elements}

\item{keep_sets}{A logical value if you wan to keep old sets}
}
\value{
A \code{TidySet} object.
}
\description{
Elements in a set not present in the other set.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = TidySet,set_in = characterORfactor,not_in = characterORfactor}: Elements present in sets but not in other sets
}}

\examples{
relations <- data.frame(sets = c("a", "a", "b", "b", "c", "c"),
                        elements = letters[seq_len(6)],
                        fuzzy = runif(6))
a <- tidySet(relations)
subtract(a, "a", "b")
subtract(a, "a", "b", keep = FALSE)
}
\seealso{
Other complements: \code{\link{complement_element}},
  \code{\link{complement_set}}
}
\concept{complements}
