% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TidySet-class}
\alias{TidySet-class}
\title{A tidy class to represent a set}
\description{
A set is a group of unique elements it can be either a fuzzy set, where the
relationship is between 0 or 1 or nominal
}
\section{Slots}{

\describe{
\item{\code{elements}}{A group of unique elements}
}}

