% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/union.R
\docType{methods}
\name{union}
\alias{union}
\alias{union,TidySet,character,character,character-method}
\title{Merge two sets}
\usage{
union(object, set1, set2, setName, ...)

\S4method{union}{TidySet,character,character,character}(object, set1, set2,
  setName, ...)
}
\arguments{
\item{object}{A TidySet object}

\item{set1, set2}{The name of the sets to be joined}

\item{setName}{The name of the new set}

\item{...}{Other arguments}
}
\value{
A \code{TidySet} object
}
\description{
Given two TidySets creates a new one merging both of them
}
\section{Methods (by class)}{
\itemize{
\item \code{object = TidySet,set1 = character,set2 = character,setName = character}: Merge two sets
}}

