% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{activate}
\alias{activate}
\alias{active}
\title{Determine the context of subsequent manipulations}
\usage{
activate(.data, what)

active(x)
}
\description{
Determine the context of subsequent manipulations
}
