% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/add_set.R
\docType{methods}
\name{add_set}
\alias{add_set}
\alias{add_set,TidySet,characterORfactor,characterORfactor,numeric-method}
\alias{add_set,TidySet,characterORfactor,characterORfactor,missing-method}
\title{Add a new set}
\usage{
add_set(object, elements, name, fuzzy, ...)


  \S4method{add_set}{TidySet,characterORfactor,characterORfactor,numeric}(object,
  elements, name, fuzzy = 1)


  \S4method{add_set}{TidySet,characterORfactor,characterORfactor,missing}(object,
  elements, name)
}
\arguments{
\item{object}{A TidySet object}

\item{elements}{The elements of the sets.}

\item{name}{The name of the new set.}

\item{fuzzy}{The membership of each element with the set.}

\item{...}{Other arguments.}
}
\value{
A \code{TidySet} object.
}
\description{
Given a TidySets adds a new sets
}
\section{Methods (by class)}{
\itemize{
\item \code{object = TidySet,elements = characterORfactor,name = characterORfactor,fuzzy = numeric}: Adds a set

\item \code{object = TidySet,elements = characterORfactor,name = characterORfactor,fuzzy = missing}: Adds a set
}}

\examples{
relations <- data.frame(sets = c(rep("a", 5), "b"),
                        elements = letters[seq_len(6)],
                        fuzzy = runif(6))
a <- tidySet(relations)
add_set(a, "a", "C")
}
\seealso{
Other add functions: \code{\link{add_relation}}

Other methods that create new sets: \code{\link{complement_element}},
  \code{\link{complement_set}}, \code{\link{intersection}},
  \code{\link{subtract}}, \code{\link{union}}

Other methods: \code{\link{TidySet-class}},
  \code{\link{activate}}, \code{\link{add_column}},
  \code{\link{add_relation}},
  \code{\link{arrange.TidySet}},
  \code{\link{complement_element}},
  \code{\link{complement_set}}, \code{\link{element_size}},
  \code{\link{elements}}, \code{\link{filter.TidySet}},
  \code{\link{group}}, \code{\link{incidence}},
  \code{\link{intersection}}, \code{\link{is.fuzzy}},
  \code{\link{move_to}}, \code{\link{mutate.TidySet}},
  \code{\link{nElements}}, \code{\link{nRelations}},
  \code{\link{nSets}}, \code{\link{name_elements<-}},
  \code{\link{name_sets<-}}, \code{\link{name_sets}},
  \code{\link{pull.TidySet}}, \code{\link{relations}},
  \code{\link{remove_column}},
  \code{\link{remove_element}},
  \code{\link{remove_relation}}, \code{\link{remove_set}},
  \code{\link{rename_elements}}, \code{\link{rename_set}},
  \code{\link{select.TidySet}}, \code{\link{set_size}},
  \code{\link{sets}}, \code{\link{subtract}},
  \code{\link{union}}
}
\concept{add functions}
\concept{methods}
\concept{methods that create new sets}
