% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/add_set.R
\docType{methods}
\name{add_set}
\alias{add_set}
\alias{add_set,TidySet,character,character-method}
\title{Add a new set}
\usage{
add_set(object, elements, setName, ...)

\S4method{add_set}{TidySet,character,character}(object, elements, setName,
  fuzzy = 1)
}
\arguments{
\item{object}{A TidySet object}

\item{elements}{The elements of the sets}

\item{setName}{The name of the new set}

\item{...}{Other arguments}
}
\value{
A \code{TidySet} object
}
\description{
Given a TidySets merges two sets into the new one.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = TidySet,elements = character,setName = character}: Adds a set
}}

