% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{adjacency}
\alias{adjacency}
\alias{adjacency_element}
\alias{adjacency_set}
\alias{adjacency.TidySet}
\title{Adjacency}
\usage{
\method{adjacency}{TidySet}(object)

adjacency_element(object)

adjacency_set(object)

\method{adjacency}{TidySet}(object)
}
\arguments{
\item{object}{A TidySet object}
}
\value{
A square matrix, 1 if two nodes are connected, 0 otherwise.
}
\description{
Are two elements connected ?
}
\examples{
x <- list("SET1" = letters[1:5], "SET2" = LETTERS[3:7])
a <- tidySet(x)
adjacency_element(a)
adjacency_set(a)
}
\seealso{
\code{\link{incidence}}
}
