% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{as.list.TidySet}
\alias{as.list.TidySet}
\title{Convert to list}
\usage{
\method{as.list}{TidySet}(x, ...)
}
\arguments{
\item{x}{A TidySet object to be coerced to a list.}

\item{...}{Placeholder for other arguments that could be passed to the
method. Currently not used.}
}
\value{
A list.
}
\description{
Converts a TidySet to a list.
}
\examples{
r <- data.frame(sets = c("A", "A", "A", "B", "C"),
             elements = c(letters[1:3], letters[2:3]),
             fuzzy = runif(5),
             info = rep_len(c("important", "very important"), 5))
TS <- tidySet(r)
TS
as.list(TS)
}
