% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basesets-package.R
\docType{package}
\name{BaseSet-package}
\alias{BaseSet}
\alias{BaseSet-package}
\title{BaseSet: Provides classes for working with sets}
\description{
A set collection, while not "tidy" in itself, can
    be thought of as three tidy data frames describing sets, elements and
    relations respectively. 'BaseSet' provides an approach to manipulate,
    load and use these virtual data frames.
}
\details{
It provides a class \code{\link{TidySets}} with lots of methods to do operations with sets
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/llrs/BaseSet}
  \item Report bugs at \url{https://github.com/llrs/BaseSet/issues}
}

}
\author{
\strong{Maintainer}: Lluís  Revilla Sancho \email{lluis.revilla@gmail.com}

}
\keyword{internal}
