% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basesets-package.R
\docType{package}
\name{BaseSet-package}
\alias{BaseSet}
\alias{BaseSet-package}
\title{BaseSet: Working with sets the tidy way}
\description{
The package implements a class and methods to work with sets, 
    doing intersection, union, complementary, power sets, cartesian product and 
    other set operations in a "tidy" way. Both for classical sets and fuzzy 
    sets. It can import sets from GMT, GAF and the OBO formats.
}
\details{
It provides a class \code{\link{TidySet}} with methods to do operations with sets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/llrs/BaseSet}
  \item Report bugs at \url{https://github.com/llrs/BaseSet/issues}
}

}
\author{
\strong{Maintainer}: Lluís Revilla Sancho \email{lluis.revilla@gmail.com} (\href{https://orcid.org/0000-0001-9747-2570}{ORCID})

Other contributors:
\itemize{
  \item Zebulun Arendsee [reviewer]
  \item Jennifer Chang [reviewer]
}

}
