% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/complement.R
\docType{methods}
\name{complement_element}
\alias{complement_element}
\alias{complement_element,TidySet,characterORfactor,character-method}
\title{Complement of elements}
\usage{
complement_element(object, elements, name, ...)

\S4method{complement_element}{TidySet,characterORfactor,character}(object,
  elements, name, keep = TRUE, keep_relations = keep,
  keep_elements = keep, keep_sets = keep)
}
\arguments{
\item{object}{A TidySet object.}

\item{elements}{The set to look for the complement.}

\item{name}{Name of the new set. By default it adds a "C_".}

\item{...}{Other arguments.}

\item{keep}{Logical value to keep all the other sets.}

\item{keep_relations}{A logical value if you wan to keep old relations}

\item{keep_elements}{A logical value if you wan to keep old elements}

\item{keep_sets}{A logical value if you wan to keep old sets}
}
\value{
A \code{TidySet} object.
}
\description{
Return the objects without the elements listed
}
\section{Methods (by class)}{
\itemize{
\item \code{object = TidySet,elements = characterORfactor,name = character}: Complement of the elements.
}}

\examples{
relations <- data.frame(sets = c("a", "a", "b", "b", "c", "c"),
                        elements = letters[seq_len(6)],
                        fuzzy = runif(6))
a <- tidySet(relations)
complement_element(a, "a", "C_a")
complement_element(a, "a", "C_a", keep = FALSE)
}
\seealso{
Other complements: \code{\link{complement_set}}
}
\concept{complements}
