% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/complement.R
\docType{methods}
\name{complement_set}
\alias{complement_set}
\alias{complement_set,TidySet,characterORfactor-method}
\title{Complement of a set}
\usage{
complement_set(object, sets, ...)

\S4method{complement_set}{TidySet,characterORfactor}(object, sets)
}
\arguments{
\item{object}{A TidySet object.}

\item{sets}{The set to look for the complement.}

\item{...}{Other arguments.}
}
\value{
A \code{TidySet} object.
}
\description{
Return the complement for a set
}
\section{Methods (by class)}{
\itemize{
\item \code{object = TidySet,sets = characterORfactor}: Complement of the sets.
}}

\examples{
relations <- data.frame(sets = c("a", "a", "b", "b", "c", "c"),
                        elements = letters[seq_len(6)],
                        fuzzy = runif(6))
a <- tidySet(relations)
complement_set(a, "a")
}
\seealso{
\code{\link{filter}}

Other complements: \code{\link{complement_element}}
}
\concept{complements}
