% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/length.R
\name{element_size}
\alias{element_size}
\alias{element_size,TidySet-method}
\title{Calculates the size of the elements}
\usage{
element_size(object, elements = NULL)

\S4method{element_size}{TidySet}(object, elements = NULL)
}
\arguments{
\item{object}{A TidySet object.}

\item{elements}{The element from which the length is calculated.}
}
\value{
A list with the size of the elements or the probability of having
that size.
}
\description{
Assuming that the fuzzy values are probabilities, calculates the probability
of being of different sizes for a given set.
}
\section{Methods (by class)}{
\itemize{
\item \code{TidySet}: Calculates the number of sets an element appears
with \code{\link[=length_set]{length_set()}}
}}

\examples{
relations <- data.frame(
    sets = c(rep("A", 5), "B", "C"),
    elements = c(letters[seq_len(6)], letters[6]),
    fuzzy = runif(7)
)
a <- tidySet(relations)
element_size(a)
}
\seealso{
cardinality

Other sizes: 
\code{\link{set_size}()}

Other methods: 
\code{\link{TidySet-class}},
\code{\link{activate}()},
\code{\link{add_column}()},
\code{\link{add_relation}()},
\code{\link{arrange.TidySet}()},
\code{\link{cartesian}()},
\code{\link{complement_element}()},
\code{\link{complement_set}()},
\code{\link{complement}()},
\code{\link{elements}()},
\code{\link{filter.TidySet}()},
\code{\link{group_by.TidySet}()},
\code{\link{group}()},
\code{\link{incidence}()},
\code{\link{intersection}()},
\code{\link{is.fuzzy}()},
\code{\link{is_nested}()},
\code{\link{move_to}()},
\code{\link{mutate.TidySet}()},
\code{\link{nElements}()},
\code{\link{nRelations}()},
\code{\link{nSets}()},
\code{\link{name_elements<-}()},
\code{\link{name_sets<-}()},
\code{\link{name_sets}()},
\code{\link{power_set}()},
\code{\link{pull.TidySet}()},
\code{\link{relations}()},
\code{\link{remove_column}()},
\code{\link{remove_element}()},
\code{\link{remove_relation}()},
\code{\link{remove_set}()},
\code{\link{rename_elements}()},
\code{\link{rename_set}()},
\code{\link{select.TidySet}()},
\code{\link{set_size}()},
\code{\link{sets}()},
\code{\link{subtract}()},
\code{\link{union}()}
}
\concept{methods}
\concept{sizes}
