% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/elements.R
\docType{methods}
\name{elements}
\alias{elements}
\alias{elements<-}
\alias{elements,TidySet-method}
\alias{elements<-,TidySet-method}
\title{Retrieve the elements}
\usage{
elements(object)

\S4method{elements}{TidySet}(object)

\S4method{elements}{TidySet}(object) <- value
}
\arguments{
\item{object}{A TidySet object.}

\item{value}{Modification of the elements.}
}
\value{
A \code{data.frame} with information about the elements
}
\description{
Given TidySet retrieve the elements.
}
\section{Methods (by class)}{
\itemize{
\item \code{TidySet}: Retrive the elements

\item \code{TidySet}: Modify the elements
}}

\seealso{
\code{\link{nElements}}

Other slots: \code{\link{relations}}, \code{\link{sets}}
}
\concept{slots}
