% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set.R, R/setcollection.R
\docType{methods}
\name{elements}
\alias{elements}
\alias{elements,Set-method}
\alias{elements,SetCollection-method}
\title{Retrieve the elements}
\usage{
elements(object)

\S4method{elements}{Set}(object)

\S4method{elements}{SetCollection}(object)
}
\arguments{
\item{object}{Either a \code{Set} or a \code{SetCollection} object}
}
\description{
Given a Set or a SetCollection retrieve the elements
}
\section{Methods (by class)}{
\itemize{
\item \code{Set}: For \code{Set} objects

\item \code{SetCollection}: For \code{SetCollection} objects
}}

