% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMT.R
\name{getGMT}
\alias{getGMT}
\title{Import GMT files}
\usage{
getGMT(con, sep = "\\t", ...)
}
\arguments{
\item{con}{File name of the GMT file}

\item{sep}{Separation of the file}

\item{...}{Other arguments passed to strsplit}
}
\value{
A TidySet object
}
\description{
Import GMT files
}
