% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/intersection.R
\docType{methods}
\name{intersection}
\alias{intersection}
\alias{intersection,TidySet,character,character,character-method}
\title{Intersection of two sets}
\usage{
intersection(object, set1, set2, setName, ...)

\S4method{intersection}{TidySet,character,character,character}(object,
  set1, set2, setName, FUN = "min", keep = FALSE)
}
\arguments{
\item{object}{A TidySet object}

\item{set1}{The name of the sets to be used}

\item{set2}{The name of the sets to be used}

\item{setName}{The name of the new set}

\item{...}{Other arguments}

\item{FUN}{A function to be applied when performing the union.
The standard intersection is the "min" function, but you can provide any other
function that given a numeric vector returns a single number.}

\item{keep}{A logical value if you want to keep originals sets.}
}
\value{
A \code{TidySet} object
}
\description{
Given a TidySets creates a new set with the elements on the both of them.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = TidySet,set1 = character,set2 = character,setName = character}: Applies the standard intersection
}}

