% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/relations.R
\docType{methods}
\name{is.fuzzy}
\alias{is.fuzzy}
\alias{is.fuzzy,TidySet-method}
\title{Fuzzy Sets}
\usage{
is.fuzzy(object)

\S4method{is.fuzzy}{TidySet}(object)
}
\arguments{
\item{object}{Object to be coerced or tested.}
}
\value{
A logical value.
}
\description{
Check if there are fuzzy sets.
}
\section{Methods (by class)}{
\itemize{
\item \code{TidySet}: Check if it is fuzzy
}}

