% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/relations.R
\docType{methods}
\name{is.fuzzy}
\alias{is.fuzzy}
\alias{is.fuzzy,TidySet-method}
\title{Fuzzy Sets}
\usage{
is.fuzzy(object)

\S4method{is.fuzzy}{TidySet}(object)
}
\arguments{
\item{object}{Object to be coerced or tested.}
}
\value{
A logical value.
}
\description{
Check if there are fuzzy sets.
}
\section{Methods (by class)}{
\itemize{
\item \code{TidySet}: Check if it is fuzzy
}}

\seealso{
Other methods: \code{\link{TidySet-class}},
  \code{\link{activate}}, \code{\link{add_column}},
  \code{\link{add_relation}},
  \code{\link{arrange.TidySet}}, \code{\link{cartesian}},
  \code{\link{complement_element}},
  \code{\link{complement_set}}, \code{\link{element_size}},
  \code{\link{elements}}, \code{\link{filter.TidySet}},
  \code{\link{group_by.TidySet}}, \code{\link{group}},
  \code{\link{incidence}}, \code{\link{intersection}},
  \code{\link{move_to}}, \code{\link{mutate.TidySet}},
  \code{\link{nElements}}, \code{\link{nRelations}},
  \code{\link{nSets}}, \code{\link{name_elements<-}},
  \code{\link{name_sets<-}}, \code{\link{name_sets}},
  \code{\link{power_set}}, \code{\link{pull.TidySet}},
  \code{\link{relations}}, \code{\link{remove_column}},
  \code{\link{remove_element}},
  \code{\link{remove_relation}}, \code{\link{remove_set}},
  \code{\link{rename_elements}}, \code{\link{rename_set}},
  \code{\link{select.TidySet}}, \code{\link{set_size}},
  \code{\link{sets}}, \code{\link{subtract}},
  \code{\link{union}}
}
\concept{methods}
