% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{length_probability}
\alias{length_probability}
\title{Calculates the probability of a single length}
\usage{
length_probability(p, n)
}
\arguments{
\item{p}{Probabilities}

\item{n}{Size}
}
\value{
A numeric value of the probability of the given size
}
\description{
Creates all the possibilities and then add them up.
}
\keyword{internal}
