% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{length_set}
\alias{length_set}
\title{Calculates the probability}
\usage{
length_set(probability)
}
\arguments{
\item{probability}{A numeric vector of probabilities.}
}
\value{
A vector with the probability of each set.
}
\description{
Given several probabilities it looks for how probable is to have a vector of
each length
}
\examples{
length_set(c(0.5, 0.1, 0.3, 0.5, 0.25, 0.23))
}
\seealso{
\code{\link[=length_probability]{length_probability()}} to calculate the probability of a specific
length.
}
