% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{multiply_probabilities}
\alias{multiply_probabilities}
\title{Calculates the probabilities}
\usage{
multiply_probabilities(p, i)
}
\arguments{
\item{p}{Probabilities}

\item{i}{Index of the complementary probability}
}
\value{
The log10 of the probability
}
\description{
Calculates the probability that all happened simultaneously
}
\keyword{internal}
