% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{multiply_probabilities}
\alias{multiply_probabilities}
\alias{independent_probabilities}
\title{Probability of a vector of probabilities}
\usage{
multiply_probabilities(p, i)

independent_probabilities(p, i)
}
\arguments{
\item{p}{Numeric vector of probabilities.}

\item{i}{Numeric integer index of the complementary probability.}
}
\value{
A numeric value of the probability.
}
\description{
Calculates the probability that all probabilities happened simultaneously.
\code{independent_probabilities()} just multiply the probabilities of the index
passed.
}
\examples{
multiply_probabilities(c(0.5, 0.1, 0.3, 0.5, 0.25, 0.23), c(1, 3))
independent_probabilities(c(0.5, 0.1, 0.3, 0.5, 0.25, 0.23), c(1, 3))
}
\seealso{
\code{\link[=length_probability]{length_probability()}}
}
