% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/names.R
\name{name_elements}
\alias{name_elements}
\alias{name_elements,TidySet-method}
\alias{name_elements<-,TidySet,characterORfactor-method}
\title{Name elements}
\usage{
name_elements(object)

\S4method{name_elements}{TidySet}(object)

\S4method{name_elements}{TidySet,characterORfactor}(object) <- value
}
\arguments{
\item{object}{A TidySet object.}

\item{value}{A character with the new names for the elements.}
}
\value{
A \code{TidySet} object.
}
\description{
Retrieve the name of the elements.
}
\section{Functions}{
\itemize{
\item \code{name_elements,TidySet-method}: Name elements

\item \code{name_elements<-,TidySet,characterORfactor-method}: Rename elements
}}

\examples{
relations <- data.frame(
    sets = c(rep("a", 5), "b"),
    elements = letters[seq_len(6)],
    fuzzy = runif(6)
)
a <- tidySet(relations)
name_elements(a)
}
\seealso{
Other names: 
\code{\link{name_elements<-}()},
\code{\link{name_sets<-}()},
\code{\link{name_sets}()},
\code{\link{rename_elements}()},
\code{\link{rename_set}()}
}
\concept{names}
