% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{nElements}
\alias{nElements}
\title{Number of elements}
\usage{
nElements(object)
}
\arguments{
\item{object}{Object to be coerced or tested.}
}
\description{
Check the number of elements of the TidySet
}
