% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/elements.R
\docType{methods}
\name{nElements}
\alias{nElements}
\alias{nElements,TidySet-method}
\title{Number of elements}
\usage{
nElements(object)

\S4method{nElements}{TidySet}(object)
}
\arguments{
\item{object}{Object to be coerced or tested.}
}
\value{
A numeric value with the number of elements.
}
\description{
Check the number of elements of the TidySet.
}
\section{Methods (by class)}{
\itemize{
\item \code{TidySet}: Return the number of elements
}}

\seealso{
Other count functions: \code{\link{nRelations}},
  \code{\link{nSets}}
}
\concept{count functions}
