% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{nElements}
\alias{nElements}
\title{Number of elements}
\usage{
nElements(object, all)
}
\arguments{
\item{object}{Object to be coerced or tested.}

\item{all}{Logical value to count all elements.}
}
\value{
A numeric value with the number of elements.
}
\description{
Check the number of elements of the TidySet.
}
\examples{
TS <- tidySet(list(A = letters[1:2], B = letters[5:7]))
nElements(TS)
}
\seealso{
Other count functions: 
\code{\link{nRelations}()},
\code{\link{nSets}()}

Other methods: 
\code{\link{TidySet-class}},
\code{\link{activate}()},
\code{\link{add_column}()},
\code{\link{add_relation}()},
\code{\link{arrange.TidySet}()},
\code{\link{cartesian}()},
\code{\link{complement}()},
\code{\link{complement_element}()},
\code{\link{complement_set}()},
\code{\link{element_size}()},
\code{\link{elements}()},
\code{\link{filter.TidySet}()},
\code{\link{group}()},
\code{\link{group_by.TidySet}()},
\code{\link{incidence}()},
\code{\link{intersection}()},
\code{\link{is.fuzzy}()},
\code{\link{is_nested}()},
\code{\link{move_to}()},
\code{\link{mutate.TidySet}()},
\code{\link{nRelations}()},
\code{\link{nSets}()},
\code{\link{name_elements<-}()},
\code{\link{name_sets}()},
\code{\link{name_sets<-}()},
\code{\link{power_set}()},
\code{\link{pull.TidySet}()},
\code{\link{relations}()},
\code{\link{remove_column}()},
\code{\link{remove_element}()},
\code{\link{remove_relation}()},
\code{\link{remove_set}()},
\code{\link{rename_elements}()},
\code{\link{rename_set}()},
\code{\link{select.TidySet}()},
\code{\link{set_size}()},
\code{\link{sets}()},
\code{\link{subtract}()},
\code{\link{union}()}
}
\concept{count functions}
\concept{methods}
