% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/relations.R
\docType{methods}
\name{nRelations}
\alias{nRelations}
\alias{nRelations,TidySet-method}
\title{Number of relations}
\usage{
nRelations(object)

\S4method{nRelations}{TidySet}(object)
}
\arguments{
\item{object}{Object to be coerced or tested.}
}
\value{
A numeric value with the number of the relations.
}
\description{
Check the number of relations of the TidySet.
}
\section{Methods (by class)}{
\itemize{
\item \code{TidySet}: Return the number of relations
}}

