% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R, R/mutate.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{filter}
\alias{mutate}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}}
}}

