% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{relations}
\alias{relations}
\title{Find the relationship}
\usage{
relations(object)
}
\arguments{
\item{object}{Object to be coerced or tested.}
}
\value{
The existing relationships
}
\description{
Find the relationship
}
