% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/relations.R
\docType{methods}
\name{relations}
\alias{relations}
\alias{relations<-}
\alias{relations,TidySet-method}
\alias{relations<-,TidySet-method}
\title{Find the relationship}
\usage{
relations(object)

relations(object) <- value

\S4method{relations}{TidySet}(object)

\S4method{relations}{TidySet}(object) <- value
}
\arguments{
\item{object}{Object to be coerced or tested.}

\item{value}{Modification of the relations.}
}
\value{
A \code{data.frame} with information about the relations between
elements and sets.
}
\description{
Method to find the relationships between the elements and the sets of a given
\code{\link{TidySet}} object
}
\section{Methods (by class)}{
\itemize{
\item \code{TidySet}: Retrieve the relations

\item \code{TidySet}: Modify the relations
}}

\seealso{
\code{\link{nRelations}}

Other slots: \code{\link{elements}}, \code{\link{sets}}
}
\concept{slots}
