% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/remove.R
\docType{methods}
\name{remove_relation}
\alias{remove_relation}
\alias{remove_relation,TidySet,character,character-method}
\title{Remove a relation}
\usage{
remove_relation(object, elements, sets, ...)

\S4method{remove_relation}{TidySet,character,character}(object, elements,
  sets)
}
\arguments{
\item{object}{A TidySet object}

\item{elements}{The elements of the sets.}

\item{sets}{The name of the new set.}

\item{...}{Other arguments}
}
\value{
A \code{TidySet} object
}
\description{
Given a TidySets removes relations between elements and sets
}
\section{Methods (by class)}{
\itemize{
\item \code{object = TidySet,elements = character,sets = character}: Removes a relation between elements and sets.
}}

