% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/remove.R
\docType{methods}
\name{remove_relation}
\alias{remove_relation}
\alias{remove_relation,TidySet,characterORfactor,characterORfactor-method}
\title{Remove a relation}
\usage{
remove_relation(object, elements, sets, ...)


  \S4method{remove_relation}{TidySet,characterORfactor,characterORfactor}(object,
  elements, sets)
}
\arguments{
\item{object}{A TidySet object}

\item{elements}{The elements of the sets.}

\item{sets}{The name of the new set.}

\item{...}{Other arguments.}
}
\value{
A \code{TidySet} object.
}
\description{
Given a TidySets removes relations between elements and sets
}
\section{Methods (by class)}{
\itemize{
\item \code{object = TidySet,elements = characterORfactor,sets = characterORfactor}: Removes a relation between elements and sets.
}}

\examples{
relations <- data.frame(sets = c(rep("a", 5), "b"),
                        elements = letters[seq_len(6)],
                        fuzzy = runif(6))
a <- tidySet(relations)
remove_relation(a, "a", "a")
}
\seealso{
Other remove functions: \code{\link{remove_element}},
  \code{\link{remove_set}}

Other methods: \code{\link{TidySet-class}},
  \code{\link{activate}}, \code{\link{add_column}},
  \code{\link{add_relation}}, \code{\link{add_set}},
  \code{\link{arrange.TidySet}},
  \code{\link{complement_element}},
  \code{\link{complement_set}}, \code{\link{element_size}},
  \code{\link{elements}}, \code{\link{filter.TidySet}},
  \code{\link{group}}, \code{\link{incidence}},
  \code{\link{intersection}}, \code{\link{is.fuzzy}},
  \code{\link{move_to}}, \code{\link{mutate.TidySet}},
  \code{\link{nElements}}, \code{\link{nRelations}},
  \code{\link{nSets}}, \code{\link{name_elements<-}},
  \code{\link{name_sets<-}}, \code{\link{name_sets}},
  \code{\link{pull.TidySet}}, \code{\link{relations}},
  \code{\link{remove_column}},
  \code{\link{remove_element}}, \code{\link{remove_set}},
  \code{\link{rename_elements}}, \code{\link{rename_set}},
  \code{\link{select.TidySet}}, \code{\link{set_size}},
  \code{\link{sets}}, \code{\link{subtract}},
  \code{\link{union}}
}
\concept{methods}
\concept{remove functions}
