% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/remove.R
\docType{methods}
\name{remove_set}
\alias{remove_set}
\alias{remove_set,TidySet,character-method}
\title{Remove sets}
\usage{
remove_set(object, sets, ...)

\S4method{remove_set}{TidySet,character}(object, sets)
}
\arguments{
\item{object}{A TidySet object.}

\item{sets}{The sets to be removed.}

\item{...}{Other arguments.}
}
\value{
A \code{TidySet} object.
}
\description{
Given a TidySets remove sets and the related relations and if
required also the elements
}
\section{Methods (by class)}{
\itemize{
\item \code{object = TidySet,sets = character}: Removes everything related to a set
}}

\examples{
relations <- data.frame(sets = c("a", "a", "b", "b", "c", "c"),
                        elements = letters[seq_len(6)],
                        fuzzy = runif(6))
a <- tidySet(relations)
remove_set(a, "b")
}
