% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/rename.R
\docType{methods}
\name{rename_set}
\alias{rename_set}
\alias{rename_set,TidySet-method}
\title{Rename sets}
\usage{
rename_set(object, old, new)

\S4method{rename_set}{TidySet}(object, old, new)
}
\arguments{
\item{object}{A TidySet object.}

\item{old}{A character vector of to be renamed.}

\item{new}{A character vector of with new names.}
}
\value{
A \code{TidySet} object.
}
\description{
Change the default names of sets and elements.
}
\section{Methods (by class)}{
\itemize{
\item \code{TidySet}: Rename sets
}}

\seealso{
\code{\link{name_sets}}

Other renames: \code{\link{rename_elements}}

Other names: \code{\link{name_elements<-}},
  \code{\link{name_elements}}, \code{\link{name_sets<-}},
  \code{\link{name_sets}}, \code{\link{rename_elements}}
}
\concept{names}
\concept{renames}
