% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Set-class}
\alias{Set-class}
\title{An S4 class to represent a set account.}
\description{
An S4 class to represent a set account.
}
\section{Slots}{

\describe{
\item{\code{elements}}{A group of unique elements}
}}

