% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/length.R
\docType{methods}
\name{set_size}
\alias{set_size}
\alias{set_size,TidySet-method}
\title{Calculates the size of the set}
\usage{
set_size(object, set = NULL)

\S4method{set_size}{TidySet}(object, set = NULL)
}
\arguments{
\item{object}{A TidySet object}

\item{set}{The sets from which the length is calculated}
}
\value{
A list with the size of the set or the probability of having that
size.
}
\description{
Calculate the probability of being of different sizes for a given set
}
\section{Methods (by class)}{
\itemize{
\item \code{TidySet}: Calculates the size of a set either fuzzy or not
}}

