% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set.R
\docType{methods}
\name{sets}
\alias{sets}
\alias{sets<-}
\alias{sets,TidySet-method}
\alias{sets<-,TidySet-method}
\alias{nSets,TidySet-method}
\title{Retrieve the names of the sets}
\usage{
sets(object)

sets(object) <- value

\S4method{sets}{TidySet}(object)

\S4method{sets}{TidySet}(object) <- value

\S4method{nSets}{TidySet}(object)
}
\arguments{
\item{object}{A \code{SetCollection} object.}

\item{value}{Modification of the sets.}
}
\value{
A \code{data.frame} with information from the sets.
}
\description{
Given a TidySet retrieve the names of the sets.
}
\section{Methods (by class)}{
\itemize{
\item \code{TidySet}: Retrieve the sets information

\item \code{TidySet}: Modify the sets information

\item \code{TidySet}: Return the number of sets
}}

\seealso{
\code{\link{nSets}}

Other slots: \code{\link{elements}},
  \code{\link{relations}}
}
\concept{slots}
