% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setcollection.R
\docType{methods}
\name{sets}
\alias{sets}
\alias{sets,SetCollection-method}
\title{Retrieve the names of the sets}
\usage{
sets(object)

\S4method{sets}{SetCollection}(object)
}
\arguments{
\item{object}{A \code{SetCollection} object}
}
\description{
Given a SetCollection retrieve the names of the sets
}
\section{Methods (by class)}{
\itemize{
\item \code{SetCollection}: For \code{SetCollection} objects
}}

