% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{size}
\alias{size}
\title{Size}
\usage{
size(object, ...)
}
\arguments{
\item{object}{A TidySet object}

\item{...}{Other arguments to filter which size should be shown.}
}
\value{
The size of the active slot. Not available for relations.
}
\description{
Calculate the size of the elements or sets
}
\examples{
relations <- data.frame(sets = c(rep("a", 5), "b", "c"),
                        elements = c(letters[seq_len(6)], letters[6]),
                        fuzzy = runif(7))
a <- tidySet(relations)
a <- activate(a, "elements")
size(a)
a <- activate(a, "sets")
size(a)
}
