% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetCollection.R
\name{tidy}
\alias{tidy}
\alias{tidy.GeneSetCollection}
\alias{tidy.GeneSet}
\title{Convert GSEABase classes to a TidySet}
\usage{
tidy(object)

\method{tidy}{GeneSetCollection}(object)

\method{tidy}{GeneSet}(object)
}
\arguments{
\item{object}{A GeneSetCollection or a GeneSet derived object}
}
\value{
A TidySet object
}
\description{
Convert GSEABase classes to a TidySet
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Converts to a tidySet given a GeneSetCollection

\item \code{GeneSet}: Converts to a tidySet given a GeneSet
}}

