% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{tidy}
\alias{tidy}
\title{tidy}
\usage{
tidy(object)
}
\arguments{
\item{object}{Object to be coerced or tested.}
}
\value{
A data.frame
}
\description{
Transform the object to a tidy/long \code{data.frame}
}
