% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-set.R
\name{TidySet}
\alias{TidySet}
\title{tidySet}
\usage{
TidySet(relations, sets = NULL, elements = NULL)
}
\arguments{
\item{relations}{(required) A data.frame with at least two columns: "sets" and "elements"}

\item{sets}{A data.frame with at least one column: "sets"}

\item{elements}{A data.frame with at least one column: "elements"}
}
\value{
A TidySet object
}
\description{
tidySet
}
\examples{
relations <- data.frame(sets = c(rep("a", 5), "b"),
elements = letters[seq_len(6)])
TidySet(relations = relations)
}
