% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-set.R
\name{tidySet}
\alias{tidySet}
\alias{tidySet.data.frame}
\alias{tidySet.list}
\alias{tidySet.matrix}
\title{Create a TidySet object}
\usage{
tidySet(relations)

\method{tidySet}{data.frame}(relations)

\method{tidySet}{list}(relations)

\method{tidySet}{matrix}(relations)
}
\arguments{
\item{relations}{(required) Depending on the method, it should be a
\code{data.frame} with at least one column for the elements and one for the
sets, a \code{list} with the elements that belong to each set, a
\code{matrix} with the elements as the rownames and the sets as the column
names.}
}
\value{
A TidySet object
}
\description{
This functions help to create a \code{TidySet} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Given the relations in a data.frame

\item \code{list}: Convert to a TidySet from a list

\item \code{matrix}: Convert an incidence matrix into a TidySet
}}

\examples{
relations <- data.frame(sets = c(rep("a", 5), "b"),
                        elements = letters[seq_len(6)])
tidySet(relations)
x <- list("a" = letters[1:5], "b" = LETTERS[3:7])
a <- tidySet(x)
x <- list("a" = letters[1:5], "b" = LETTERS[3:7], "c" = runif(5))
a <- tidySet(x) # An error for mixing letters and numbers
# Numeric input should be named
x <- list("a" = c("A" = 0.1, "B" = 0.5), "b" = c("A" = 0.2, "B" = 1))
a <- tidySet(x)
tidySet.matrix(incidence(a))
}
\seealso{
\code{\link{TidySet-class}}
}
