% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetCollection.R, R/tidy-set.R
\name{tidy}
\alias{tidy}
\alias{tidy.GeneSetCollection}
\alias{tidy.GeneSet}
\alias{tidySet}
\alias{tidySet.data.frame}
\alias{tidySet.list}
\alias{tidySet.matrix}
\title{Create a TidySet object}
\usage{
tidy(object)

\method{tidy}{GeneSetCollection}(object)

\method{tidy}{GeneSet}(object)

tidySet(relations)

\method{tidySet}{data.frame}(relations)

\method{tidySet}{list}(relations)

\method{tidySet}{matrix}(relations)
}
\arguments{
\item{relations, object}{An object to be coerced to a TidySet.}
}
\value{
A TidySet object
}
\description{
This functions help to create a \code{TidySet} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Converts to a tidySet given a GeneSetCollection

\item \code{GeneSet}: Converts to a tidySet given a GeneSet

\item \code{data.frame}: Given the relations in a data.frame

\item \code{list}: Convert to a TidySet from a list

\item \code{matrix}: Convert an incidence matrix into a TidySet
}}

\examples{
relations <- data.frame(sets = c(rep("a", 5), "b"),
                        elements = letters[seq_len(6)])
tidySet(relations)
x <- list("A" = letters[1:5], "B" = LETTERS[3:7])
tidySet(x)
# A fuzzy set taken encoded as a list
A <- runif(5)
names(A) <- letters[1:5]
B <- runif(5)
names(B) <- letters[3:7]
relations <- list(A, B)
tidySet(relations)
\dontrun{
x <- list("A" = letters[1:5], "B" = LETTERS[3:7], "c" = runif(5))
a <- tidySet(x) # An error for mixing letters and numbers
}
# Numeric input should be named
x <- list("a" = c("A" = 0.1, "B" = 0.5), "b" = c("A" = 0.2, "B" = 1))
a <- tidySet(x)
tidySet(incidence(a))
}
\seealso{
\code{\link{TidySet-class}}
}
