% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{union}
\alias{union}
\title{Merge two sets}
\usage{
union(object1, object2)
}
\arguments{
\item{object1}{A TidySet object}

\item{object2}{A TidySet object}
}
\value{
A \code{TidySet} object
}
\description{
Given two TidySets creates a new one merging both of them
}
