% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/union.R
\docType{methods}
\name{union}
\alias{union}
\alias{union,TidySet,character,character,character-method}
\title{Merge two sets}
\usage{
union(object, set1, set2, setName, ...)

\S4method{union}{TidySet,character,character,character}(object, set1, set2,
  setName, FUN = "max", keep = FALSE)
}
\arguments{
\item{object}{A TidySet object}

\item{set1, set2}{The name of the sets to be used.}

\item{setName}{The name of the new set.}

\item{...}{Other arguments.}

\item{FUN}{A function to be applied when performing the union.
The standard union is the "max" function, but you can provide any other
function that given a numeric vector returns a single number.}

\item{keep}{A logical value if you want to keep originals sets.}
}
\value{
A \code{TidySet} object.
}
\description{
Given a TidySets merges two sets into the new one.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = TidySet,set1 = character,set2 = character,setName = character}: Applies the standard union
}}

